Added new JPEG2000 image viewer to use IIPimage/OpenSeaDragon.
More changes and refinements to the new item viewer.
Also, refactored many of the HTML writing methods so they have the same name from the SobekCK.aspx in-line calls all the way through the HtmlMainWriter, the HtmlSubWriters, and the Item Viewers.  Any method that writes directly to the output stream begins with 'Write_' and any that add controls to placeholders begins with 'Add_'.
Removed the nagivationMenuPlaceHolder from the SobekCM page and all subsequent pages.
Cleaned up the HtmlWriter so code is much more straightforward and uses the recently added Html behaviors from the subwriters, rather than hard coding things in the HtmlWriter.
Many more small things.
(Currently Aware_JP2_ItemViewer is still being used for serving JP2, as I work through a blurry issue in IIPimage)