Performance changes and email change
Changed all calls to Response.Redirect to ensure the ThreadAbortException is not thrown
Added new Request_Completed flag to the SobekCM_Navigation_Object (aka Current Mode) to show no more execution should occur and added checks for the flag throughout the ASP pagees and main writes/subwriters
Emails are no longer sent to the main sys admin when a bad web skin is requested. If it's a robot they receive a 404 error, otherwise just forwarded to the default web skin.