Settings structure updated
Updated the settings object significantly.  A non-static object now appears in the Core class that holds all the data and contains a DataContract for conversion to REST API.  There is now a static singleton-pattern class to access the settings object and a settings builder.  The later two objects are currently in the SobekCM_Library project.  Updated all references to the settings throughout the code.  Also, pulled the list of search stop words out of the settings, since the final UI layer should never need this list.

Access to the settings object should now be via "InstanceWide_Settings_Singleton.Settings.", rather than "SobekCM_Library_Settings."

This is all to prepare for the division between the REST API/UI, but also allows the builder to use the Settings instance, rather than a static class, which *should* correct issues of the settings object retaining the data from a previous instance when working in a multi-instance builder environment.

Work on the builder modularization is now in test.