SobekCM Engine Microservice Handler added
Tested multiple JSON creating libraries for serialization speed (and looked at published benchmarks) and chose to use the JIL library for JSON support and the protobuf-net library for support for protocol buffer messages for inter-process communications.  Updated the configuration reader to support multiple protocols.  Added the microservices configuration file to add a single endpoint, to return the information about an aggregation.  Engine can be invoked with either the /engine/ URL segment, or sobekcm.svc, which is what the custom http handler responds to.  Will continue to decorate the Item_Aggregation class and build out that endpoint and also add an endpoint for getting the basic information about a single item.